requirements.txt
cffi==1.15.0
cryptography==37.0.2
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
typing-extensions==4.2.0
